Protein acetylation feature annotation

Protein acetylation feature annotation

Format

A data frame with 18341 rows and 12 variables:

protein_id

character, RefSeq identifier in format [accession].[version], e.g., NP_001004415.1. Note: contaminants should be identified with the following nomenclature: "Contaminant_XXXX"

redundant_ids

character, pipe-separated list of additional RefSeq identifiers with redundant peptide-matching sequences to protein_id that cannot be resolved by inference due to lack of unique peptides

is_contaminant

logical, whether protein_id is a contaminant

peptide_score

double, MSGF+ SpecE Value or Spectrum Mill score (1/bestScore)

sequence

character, peptide sequence determined by MSGF+ or Spectrum Mill analysis of LC-MS/MS features

organism_name

character, organism whose protein collection database was used for LC-MS/MS peptide identification, e.g., Rattus norvegicus

ptm_id

character, concatenation of protein_id and the modified amino acid residue (e.g., serine residue, 20 denoted as S20s), with "-" as a separator

ptm_peptide

character, concatenation of ptm_id and peptide with "-" as a separator. Values in this column are unique.

confident_score

double, score indicating confidence of modified amino acid residue localization by LC-MS/MS fragmentation as calculated by Ascore algorith (0-1000) or Spectrum Mill (0-1)

confident_site

logical, whether a modified site was confidently localized by LC-MS/MS fragmentation

tissue

character, tissue abbreviation, one of TISSUE_ABBREV

assay

character, assay "ACETYL"

Source

pass1b-06/results/proteomics-untargeted/*/prot-ac/*rii-results.txt

Details

PHOSPHO feature annotation is only available via download from Google Cloud Storage: https://storage.googleapis.com/motrpac-rat-training-6mo-extdata/feature-annot/PHOSPHO_FEATURE_ANNOT.rda. You can use MotrpacRatTraining6mo::load_feature_annotation() to download and return this file.