Gene-centric universes

Nested vectors of the set of genes tested for each dataset used to specify custom backgrounds for enrichment analyses

Usage

GENE_UNIVERSES

Format

A list of lists of lists of vectors. Access a vector of measured/tested genes with GENE_UNIVERSES[[gene_identifier]][[ome]][[tissue]], where:

gene_identifier

one of "entrez_gene", "gene_symbol", "ensembl_gene", "rgd_gene"

ome

character, assay abbreviation, one of ASSAY_ABBREV

tissue

character, tissue abbreviation, one of TISSUE_ABBREV

Details

If you are performing any kind of enrichment test that does not accept the full set of measured features as the input, you should specify a custom background. For example, if you are performing enrichment for an unordered list of differential features, then the background should specify the full set of features tested during differential analysis. This object provides the gene-centric lists of "expressed" features, i.e. those tested during differential analysis, for all assay and tissue combinations.

For convenience, universes are specified with four different types of gene identifers: Entrez/NCBI IDs ("entrez_gene"), gene symbols ("gene_symbol"), Ensembl IDs ("ensembl_gene"), and RGD IDs ("rgd_gene"). This object is a list of lists, with one list per type of gene identifier. List structure is gene identifier > assay abbreviation > tissue abbreviation, e.g. GENE_UNIVERSES$entrez_gene$PHOSPHO$HEART is the unique list of Entrez genes associated with all measured phosphosites in the heart.

Universes for ATAC and METHYL correspond to the universe of expressed genes, i.e. TRNSCRPT, not actually the full set of genes associated with any features measured in these epigenetic assays.

Regardless of the gene identifier, features in METAB universe lists are always KEGG IDs, not actually gene identifiers.