Graph components

The feature sets selected using the repfdr results. Each value is a named list, where the name is the label for the edge or node, and the members are all of the features that belong to that label, in the format ASSAY_ABBREV;TISSUE_ABBREV;[feature_ID].

Usage

GRAPH_COMPONENTS

Format

List of lists:

edge_sets

named list, where the name is the label for the edge in the format [fromNode]---[toNode] and the value is a vector of features that belong to that edge

node_sets

named list, where the name is the label for the node in the format [week]w_F[state]_M[state] and the value is a vector of features that belong to that edge

Details

Nodes are named in the format [week]w_F[state]_M[state], where [week] is one of 1, 2, 4, or 8 to specify the training point, [state] takes a value of 0 (null), 1 (up-regulated), or -1 (down-regulated). Altogether, then name of a node specifies the state of a feature in each sex at a specific time point. For example, the node 1w_F-1_M-1 specifies all features that are down-regulated in females (F-1) and males (M-1) at the 1-week training time point (1w).