Differential analysis of multiplexed immunoassays

Timewise summary statistics and training FDR from differential analysis (DA) that tests the effect of training on each immunoassay analyte within each sex. One data frame per tissue.

Usage

IMMUNO_PLASMA_DA

IMMUNO_HIPPOC_DA

IMMUNO_CORTEX_DA

IMMUNO_ADRNL_DA

IMMUNO_TESTES_DA

IMMUNO_OVARY_DA

IMMUNO_SKMGN_DA

IMMUNO_BAT_DA

IMMUNO_WATSC_DA

IMMUNO_COLON_DA

IMMUNO_SMLINT_DA

IMMUNO_LIVER_DA

IMMUNO_SKMVL_DA

IMMUNO_HEART_DA

IMMUNO_KIDNEY_DA

IMMUNO_SPLEEN_DA

IMMUNO_LUNG_DA

Format

A data frame with 17 variables:

feature

character, unique feature identifier in the format 'ASSAY_ABBREV;TISSUE_ABBREV;feature_ID' only for training-regulated features at 5% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See REPEATED_FEATURES for details.

assay

character, assay abbreviation, one of ASSAY_ABBREV

assay_code

character, assay code used in data release. See MotrpacBicQC::assay_codes.

tissue

character, tissue abbreviation, one of TISSUE_ABBREV

tissue_code

character, tissue code used in data release. See MotrpacBicQC::bic_animal_tissue_code.

feature_ID

character, MoTrPAC feature identifier

dataset

character, name of LUMINEX panel, e.g., 'rat-mag27plex'

sex

character, one of 'male' or 'female'

comparison_group

character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'

p_value

double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals

adj_p_value

double, adjusted p-value from 'p_value' column. P-values are BY-adjusted across all datasets within a given assay/ome.

logFC

double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals

logFC_se

double, standard error of the log fold-change

covariates

character, comma-separated list of adjustment variables or NA

comparison_average_intensity

double, average intensity among the replicates in 'comparison_group'

reference_average_intensity

double, average intensity among the replicates in the group of sex-matched sedentary control animals

selection_fdr

double, adjusted training p-value used to select training-regulated analytes. P-values are IHW-adjusted across all datasets within a given assay with tissue as a covariate.

An object of class data.frame with 368 rows and 17 columns.

An object of class data.frame with 368 rows and 17 columns.

An object of class data.frame with 368 rows and 17 columns.

An object of class data.frame with 184 rows and 17 columns.

An object of class data.frame with 184 rows and 17 columns.

An object of class data.frame with 424 rows and 17 columns.

An object of class data.frame with 424 rows and 17 columns.

An object of class data.frame with 424 rows and 17 columns.

An object of class data.frame with 280 rows and 17 columns.

An object of class data.frame with 312 rows and 17 columns.

An object of class data.frame with 312 rows and 17 columns.

An object of class data.frame with 312 rows and 17 columns.

An object of class data.frame with 312 rows and 17 columns.

An object of class data.frame with 216 rows and 17 columns.

An object of class data.frame with 208 rows and 17 columns.

An object of class data.frame with 216 rows and 17 columns.

Details

Reproduce our analysis with MotrpacRatTraining6mo::immuno_timewise_da() and MotrpacRatTraining6mo::immuno_training_da().