Multiplexed immunoassay metadata and QC

Multiplexed immunoassay (IMMUNO) experimental and quantification QC metrics

Usage

IMMUNO_META

Format

A data frame with 1511 rows and 23 variables:

viallabel

character, sample identifier

tissue

character, tissue abbreviation, one of TISSUE_ABBREV

tissue_code

character, tissue code used in data release. See MotrpacBicQC::bic_animal_tissue_code.

bid

integer, biospecimen ID

luminex_sample_name

character, sample name used by HIMC in raw data files

panel_name

character, LUMINEX panel name

plate_id

character, plate ID in format "[date]-[tissues]_[panel_name]"

ppt_type

character, species

weight_mg

integer, sample weight in mg (solid tissue only)

volume_ul

integer, sample volume in uL (plasma only)

date_extracted

character, date of protein extraction

operators

character, initials of experimentalists

protein_conc_mg_ml

double, protein concentration in mg/mL

vol_plate1_ul

integer, volume (uL) of sample submitted in first master plate (plate used for LUMINEX assays)

min_vol_plate2_ul

integer, volume (uL) of remaining sample submitted in second master plate (backup plate)

comments

character, comments about protein extraction

sample_well_position

character, well position in plates submitted to the HIMC

luminex_well

character, well position for LUMINEX assay

sex

character, one of 'male' or 'female'

group

character, intervention group, one of "1w", "2w", "4w", "8w", "control"

log2_CHEX1

double, custom Assay CHEX control beads to monitor instrument performance, log-transformed

log2_CHEX2

double, custom Assay CHEX control beads to monitor application of the detection antibody, log-transformed

log2_CHEX3

double, custom Assay CHEX control beads to monitor application of the fluorescent reporter, log-transformed

log2_CHEX4

double, custom Assay CHEX control beads to monitor nonspecific binding, log-transformed

Details

Protein extractions from the same sample/vial label were used for multiple panels, so metadata must be matched to unique samples using both viallabel and panel_name.