Differential analysis of merged metabolomics datasets

Timewise summary statistics and training FDR from differential analysis (DA) that tests the effect of training on each metabolite within each sex. One data frame per tissue.

Usage

METAB_PLASMA_DA

METAB_HIPPOC_DA

METAB_CORTEX_DA

METAB_HYPOTH_DA

METAB_SKMGN_DA

METAB_SKMVL_DA

METAB_HEART_DA

METAB_KIDNEY_DA

METAB_ADRNL_DA

METAB_COLON_DA

METAB_SPLEEN_DA

METAB_TESTES_DA

METAB_OVARY_DA

METAB_VENACV_DA

METAB_LUNG_DA

METAB_SMLINT_DA

METAB_LIVER_DA

METAB_BAT_DA

METAB_WATSC_DA

Format

A data frame with 27 variables:

feature

character, unique feature identifier in the format 'ASSAY_ABBREV;TISSUE_ABBREV;feature_ID' only for training-regulated features at 5% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See REPEATED_FEATURES for details.

assay

character, assay abbreviation, one of ASSAY_ABBREV

assay_code

character, assay code used in data release. See MotrpacBicQC::assay_codes.

tissue

character, tissue abbreviation, one of TISSUE_ABBREV

tissue_code

character, tissue code used in data release. See MotrpacBicQC::bic_animal_tissue_code.

feature_ID

character, MoTrPAC feature identifier

dataset

character, metabolomics platform (metab-u-ionpneg, metab-u-lrpneg, metab-u-lrppos, metab-u-hilicpos, metab-u-rpneg, metab-u-rppos, metab-t-amines, metab-t-oxylipneg, metab-t-tca, metab-t-nuc, metab-t-acoa, metab-t-ka) the feature was measured in. 'meta-reg' specifies results from the metabolomics meta-regression for repeated features.

site

character, Chemical Analysis Site (CAS) name

is_targeted

logical, is this a targeted platform?

sex

character, one of 'male' or 'female'

comparison_group

character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'

p_value

double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals

adj_p_value

double, adjusted p-value from 'p_value' column. P-values are BY-adjusted across all datasets within a given assay/ome.

logFC

double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals

logFC_se

double, standard error of the log fold-change

tscore

double, t statistic

covariates

character, comma-separated list of adjustment variables or NA

comparison_average_intensity

double, average intensity among the replicates in 'comparison_group'

reference_average_intensity

double, average intensity among the replicates in the group of sex-matched sedentary control animals

metabolite_refmet

character, RefMet name of metabolite

cv

double, feature coefficient of variation in the dataset

metabolite

character, name of metabolite as appears in the CAS's data

control_cv

double, feature coefficient of variation in the dataset

mz

double, mass over charge

rt

double, retention time

neutral_mass

double, neutral mass

selection_fdr

double, adjusted training p-value used to select training-regulated analytes. P-values are IHW-adjusted across all datasets within a given assay with tissue as a covariate.

An object of class data.frame with 10000 rows and 27 columns.

An object of class data.frame with 1672 rows and 27 columns.

An object of class data.frame with 1736 rows and 27 columns.

An object of class data.frame with 9880 rows and 27 columns.

An object of class data.frame with 1584 rows and 27 columns.

An object of class data.frame with 11440 rows and 27 columns.

An object of class data.frame with 11952 rows and 27 columns.

An object of class data.frame with 1752 rows and 27 columns.

An object of class data.frame with 1888 rows and 27 columns.

An object of class data.frame with 1816 rows and 27 columns.

An object of class data.frame with 872 rows and 27 columns.

An object of class data.frame with 912 rows and 27 columns.

An object of class data.frame with 1278 rows and 27 columns.

An object of class data.frame with 11736 rows and 27 columns.

An object of class data.frame with 1800 rows and 27 columns.

An object of class data.frame with 11896 rows and 27 columns.

An object of class data.frame with 10402 rows and 27 columns.

An object of class data.frame with 10288 rows and 27 columns.

Details

Reproduce our analysis with MotrpacRatTraining6mo::metab_training_da() and MotrpacRatTraining6mo::metab_timewise_da().