Meta-regression of metabolomics differential analysis results

Timewise summary statistics and training FDR from differential analysis (DA) that tests the effect of training on each metabolite within each sex. One data frame per tissue.

For metabolites measured on more than one platform, the consensus result may be provided. For metabolites measured on a single platform, these results are identical to METAB_DA.

Usage

METAB_ADRNL_DA_METAREG

METAB_BAT_DA_METAREG

METAB_COLON_DA_METAREG

METAB_CORTEX_DA_METAREG

METAB_HEART_DA_METAREG

METAB_HIPPOC_DA_METAREG

METAB_HYPOTH_DA_METAREG

METAB_KIDNEY_DA_METAREG

METAB_LIVER_DA_METAREG

METAB_LUNG_DA_METAREG

METAB_OVARY_DA_METAREG

METAB_PLASMA_DA_METAREG

METAB_SKMGN_DA_METAREG

METAB_SKMVL_DA_METAREG

METAB_SMLINT_DA_METAREG

METAB_SPLEEN_DA_METAREG

METAB_TESTES_DA_METAREG

METAB_VENACV_DA_METAREG

METAB_WATSC_DA_METAREG

Format

A data frame with 30 variables:

feature

character, unique feature identifier in the format 'ASSAY_ABBREV;TISSUE_ABBREV;feature_ID' only for training-regulated features at 5% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See REPEATED_FEATURES for details.

assay

character, assay abbreviation, one of ASSAY_ABBREV

assay_code

character, assay code used in data release. See MotrpacBicQC::assay_codes.

tissue

character, tissue abbreviation, one of TISSUE_ABBREV

tissue_code

character, tissue code used in data release. See MotrpacBicQC::bic_animal_tissue_code.

feature_ID

character, MoTrPAC feature identifier

dataset

character, metabolomics platform (metab-u-ionpneg, metab-u-lrpneg, metab-u-lrppos, metab-u-hilicpos, metab-u-rpneg, metab-u-rppos, metab-t-amines, metab-t-oxylipneg, metab-t-tca, metab-t-nuc, metab-t-acoa, metab-t-ka) the feature was measured in. 'meta-reg' specifies results from the metabolomics meta-regression for repeated features.

site

character, Chemical Analysis Site (CAS) name

is_targeted

logical, is this a targeted platform?

sex

character, one of 'male' or 'female'

comparison_group

character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'

p_value

double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals

adj_p_value

double, adjusted p-value from 'p_value' column. P-values are BY-adjusted across all datasets within a given assay/ome.

logFC

double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals

logFC_se

double, standard error of the log fold-change

tscore

double, t statistic

zscore

double, z statistic

covariates

character, comma-separated list of adjustment variables or NA

comparison_average_intensity

double, average intensity among the replicates in 'comparison_group'

reference_average_intensity

double, average intensity among the replicates in the group of sex-matched sedentary control animals

metabolite_refmet

character, RefMet name of metabolite

cv

double, feature coefficient of variation in the dataset

metabolite

character, name of metabolite as appears in the CAS's data

control_cv

double, feature coefficient of variation in the dataset

mz

double, mass over charge

rt

double, retention time

neutral_mass

double, neutral mass

meta_reg_het_p

double, for metabolites with multiple measurements, the meta-regression heterogeneity p-value, where a smaller p-value indicates more disagreement between platforms. One value per feature (not per training group).

meta_reg_pvalue

double, for metabolites with multiple measurements, the meta-regression p-value. One value per feature (not per training group).

selection_fdr

double, adjusted training p-value used to select training-regulated analytes. P-values are IHW-adjusted across all datasets within a given assay with tissue as a covariate.

An object of class data.frame with 9086 rows and 30 columns.

An object of class data.frame with 1888 rows and 30 columns.

An object of class data.frame with 1672 rows and 30 columns.

An object of class data.frame with 9872 rows and 30 columns.

An object of class data.frame with 8920 rows and 30 columns.

An object of class data.frame with 1736 rows and 30 columns.

An object of class data.frame with 10376 rows and 30 columns.

An object of class data.frame with 10264 rows and 30 columns.

An object of class data.frame with 10208 rows and 30 columns.

An object of class data.frame with 912 rows and 30 columns.

An object of class data.frame with 8696 rows and 30 columns.

An object of class data.frame with 8472 rows and 30 columns.

An object of class data.frame with 1584 rows and 30 columns.

An object of class data.frame with 1800 rows and 30 columns.

An object of class data.frame with 1816 rows and 30 columns.

An object of class data.frame with 872 rows and 30 columns.

An object of class data.frame with 1278 rows and 30 columns.

An object of class data.frame with 8808 rows and 30 columns.

Details

TODO

Reproduce our analysis with TODO.