Differential analysis results of RRBS data

Timewise summary statistics and training FDR from differential analysis (DA) that tests the effect of training on each methylation feature within each sex. One data frame per tissue.

Format

A data frame with 20 variables:

feature

character, unique feature identifier in the format 'ASSAY_ABBREV;TISSUE_ABBREV;feature_ID' only for training-regulated features at 5% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See REPEATED_FEATURES for details.

assay

character, assay abbreviation, one of ASSAY_ABBREV

assay_code

character, assay code used in data release. See MotrpacBicQC::assay_codes.

tissue

character, tissue abbreviation, one of TISSUE_ABBREV

tissue_code

character, tissue code used in data release. See MotrpacBicQC::bic_animal_tissue_code.

feature_ID

character, MoTrPAC feature identifier

sex

character, one of 'male' or 'female'

comparison_group

character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'

p_value

double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals

adj_p_value

double, adjusted p-value from 'p_value' column. P-values are BY-adjusted across all datasets within a given assay/ome.

logFC

double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals

zscore

double, z statistic

covariates

character, comma-separated list of adjustment variables or NA

removed_samples

character, comma-separated list of vial labels excluded from differential analysis or NA

Chr

integer, chromosome

Locus

character, base pair range of feature

EntrezID

character, Entrez ID of closest gene

Symbol

character, gene symbol of closest gene

fscore

numeric, F-score

selection_fdr

double, adjusted training p-value used to select training-regulated analytes. P-values are IHW-adjusted across all datasets within a given assay with tissue as a covariate.

Details

Unfiltered METHYL differential analysis results are only available via download from Google Cloud Storage. For example, https://storage.googleapis.com/motrpac-rat-training-6mo-extdata/METHYL_BAT_DA.rda is the file for brown adipose tissue (BAT) results. You can change the name of the file to specify other tissues including: HEART, HIPPOC, KIDNEY, LIVER, LUNG, SKMGN (gastrocnemius skeletal muscle), and WATSC (subcutaneous white adipose tissue). You can also use MotrpacRatTraining6mo::get_rdata_from_url() to download differential analysis results for ATAC and METHYL. For more details about these files see the readme of this repository at https://github.com/MoTrPAC/MotrpacRatTraining6mo/blob/main/README.md.

The differential analysis results in this data frame were computed using edgeR::glmQLFTest() after removing a few outlier low quality samples from the kidney dataset.

For training-regulated METHYL features at 5% FDR, see TRAINING_REGULATED_FEATURES.

Reproduce our analysis with MotrpacRatTraining6mo::rrbs_differential_analysis().