MotrpacRatTraining6moData: Data for analysis of the MoTrPAC endurance exercise training study in 6-month-old rats

The goal of the Molecular Transducers of Physical Activity Consortium (MoTrPAC) is to study the molecular changes that occur during and after exercise and ultimately advance the understanding of how physical activity improves and preserves health. This package provides convenient access to the multi-tissue, multi-omic data generated by MoTrPAC's endurance exercise training study in 6-month-old rats.

See also

Author

Maintainer: Nicole Gay nicole.r.gay@gmail.com

Authors:

Pierre Jean Beltran pjeanbeltran@gmail.com

David Amar ddam.am@gmail.com

David Jimenez-Morales davidjm@stanford.edu

MoTrPAC Study Group [data contributor]

Other contributors:

Byron Jaeger bjaeger@wakehealth.edu [reviewer]

NIH Common Fund [funder]

National Institute of Diabetes and Digestive and Kidney Diseases (NIDDK) [funder]

National Institute of Arthritis and Musculoskeletal Diseases (NIAMS) [funder]

National Institute on Aging (NIA) [funder]

National Institute of Biomedical Imaging and Bioengineering [funder]