Phenotypic data

Phenotypic data for samples from the MoTrPAC endurance exercise training study in 6-month-old rats. One row per sample (viallabel).

Usage

PHENO

Format

A data frame with 5955 rows and 510 variables:

pid

integer, unique, randomly generated 8 digit numeric identifier used in linkage to phenotypic data

bid

integer, unique, randomly generated 8 digit numeric identifier used in linkage to phenotypic data

labelid

double, unique 11 digit identifier for specimen label ID, originating at the collection site, that provides a link to specimen processing and used for shipments to the biorepository (same as viallabel only in instances where aliquots were not further processed at the biorepository)

viallabel

character, unique 11 digit numeric identifier of a sample vial made up of bid, key.sacrificetime, specimen.processing.sampletypedescription, and aliquot number

key.protocol

character, protocol in which the procedures for sample collection were outlined

key.agegroup

character, age group for the animal

key.d_arrive

character, date animal arrived at site (same as registration.d_arrive)

key.d_sacrifice

character, date animal was sacrificed (same as specimen.collection.d_visit)

key.intervention

character, exercise group defined by those who complete exercise training

key.sacrificetime

character, weeks since the beginning of the intervention

key.anirandgroup

character, description of the randomization group

key.siteid

character, site where the samples were collected

registration.d_visit

character, date registration was completed

registration.days_visit

integer, day registration was completed relative to registration.d_arrive (count)

registration.staffid

integer, registration staff ID

registration.siteid

character, registration site

registration.formname

character, registration form name

registration.versionnbr

integer, registration form version number

registration.ratid

character, site rat ID

registration.d_arrive

character, date the rat arrived at the site

registration.days_arrive

integer, day of arrival relative to registration.d_arrive (count)

registration.batchnumber

integer, registration batch number

registration.d_reverselight

character, date the rat was housed in reverse light

registration.days_reverselight

integer, day the rat was housed in reverse light relative to registration.d_arrive (count)

registration.d_birth

character, rat date of birth (+/- two weeks)

registration.days_birth

integer, rat day of birth relative to registration.d_arrive (count)

registration.sex

character, sex, one of "male", "female"

registration.weight

double, rat's body weight

registration.cagenumber

integer, cage number in which the rat was housed

familiarization.d_visit

character, date familiarization was completed

familiarization.days_visit

integer, day familiarization was completed relative to registration.d_arrive (count)

familiarization.staffid

integer, familiarization staff ID

familiarization.siteid

character, familiarization site

familiarization.formname

character, familiarization form name

familiarization.versionnbr

integer, familiarization form version number

familiarization.d_treadmillbegin

character, date the rat began treadmill familiarization

familiarization.days_treadmillbegin

integer, day the rat began treadmill familiarization relative to registration.d_arrive (count)

familiarization.d_treadmillcomplete

character, date the rat completed treadmill familiarization

familiarization.days_treadmillcomplete

integer, day the rat completed treadmill familiarization relative to registration.d_arrive (count)

familiarization.activity_score

double, rat's activity score

familiarization.compliant

character, was this rat compliant?

familiarization.weight

double, rat's body weight

training.d_visit

character, date intervention started

training.days_visit

integer, day intervention started relative to registration.d_arrive

training.staffid

integer, training staff ID

training.siteid

character, training site

training.formname

character, training form name

training.versionnbr

integer, training form version number

training.day1date

character, date of given training day

training.day1_days

integer, day of given training day relative to registration.d_arrive (count)

training.day1time

character, time of given training day

training.day1_treadmillspeed

double, treadmill speed on given training day

training.day1_treadmillincline

double, treadmill incline on given training day

training.day1_timeontreadmill

integer, time on treadmill on given training day

training.day1_weight

double, weight on given training day

training.day1_posttrainlact

double, post-training lactate on given training day

training.day1_score

integer, score on given training day

training.day1_comments

character, comments on given training day

training.day2date

character, date of given training day

training.day2_days

integer, day of given training day relative to registration.d_arrive (count)

training.day2time

character, time of given training day

training.day2_treadmillspeed

double, treadmill speed on given training day

training.day2_treadmillincline

double, treadmill incline on given training day

training.day2_timeontreadmill

integer, time on treadmill on given training day

training.day2_score

integer, score on given training day

training.day2_comments

character, comments on given training day

training.day3date

character, date of given training day

training.day3_days

integer, day of given training day relative to registration.d_arrive (count)

training.day3time

character, time of given training day

training.day3_treadmillspeed

double, treadmill speed on given training day

training.day3_treadmillincline

double, treadmill incline on given training day

training.day3_timeontreadmill

integer, time on treadmill on given training day

training.day3_score

integer, score on given training day

training.day3_comments

character, comments on given training day

training.day4date

character, date of given training day

training.day4_days

integer, day of given training day relative to registration.d_arrive (count)

training.day4time

character, time of given training day

training.day4_treadmillspeed

double, treadmill speed on given training day

training.day4_treadmillincline

double, treadmill incline on given training day

training.day4_timeontreadmill

integer, time on treadmill on given training day

training.day4_score

integer, score on given training day

training.day4_comments

character, comments on given training day

training.day5date

character, date of given training day

training.day5_days

integer, day of given training day relative to registration.d_arrive (count)

training.day5time

character, time of given training day

training.day5_treadmillspeed

double, treadmill speed on given training day

training.day5_treadmillincline

double, treadmill incline on given training day

training.day5_timeontreadmill

integer, time on treadmill on given training day

training.day5_posttrainlact

double, post-training lactate on given training day

training.day5_score

integer, score on given training day

training.day5_comments

character, comments on given training day

training.day6date

character, date of given training day

training.day6_days

integer, day of given training day relative to registration.d_arrive (count)

training.day6time

character, time of given training day

training.day6_treadmillspeed

double, treadmill speed on given training day

training.day6_treadmillincline

double, treadmill incline on given training day

training.day6_timeontreadmill

integer, time on treadmill on given training day

training.day6_weight

double, weight on given training day

training.day6_posttrainlact

double, post-training lactate on given training day

training.day6_score

integer, score on given training day

training.day6_comments

character, comments on given training day

training.day7date

character, date of given training day

training.day7_days

integer, day of given training day relative to registration.d_arrive (count)

training.day7time

character, time of given training day

training.day7_treadmillspeed

double, treadmill speed on given training day

training.day7_treadmillincline

double, treadmill incline on given training day

training.day7_timeontreadmill

integer, time on treadmill on given training day

training.day7_posttrainlact

double, post-training lactate on given training day

training.day7_score

integer, score on given training day

training.day7_comments

character, comments on given training day

training.day8date

character, date of given training day

training.day8_days

integer, day of given training day relative to registration.d_arrive (count)

training.day8time

character, time of given training day

training.day8_treadmillspeed

double, treadmill speed on given training day

training.day8_treadmillincline

double, treadmill incline on given training day

training.day8_timeontreadmill

integer, time on treadmill on given training day

training.day8_posttrainlact

double, post-training lactate on given training day

training.day8_score

integer, score on given training day

training.day8_comments

character, comments on given training day

training.day9date

character, date of given training day

training.day9_days

integer, day of given training day relative to registration.d_arrive (count)

training.day9time

character, time of given training day

training.day9_treadmillspeed

double, treadmill speed on given training day

training.day9_treadmillincline

double, treadmill incline on given training day

training.day9_timeontreadmill

integer, time on treadmill on given training day

training.day9_posttrainlact

double, post-training lactate on given training day

training.day9_score

integer, score on given training day

training.day9_comments

character, comments on given training day

training.day10date

character, date of given training day

training.day10_days

integer, day of given training day relative to registration.d_arrive (count)

training.day10time

character, time of given training day

training.day10_treadmillspeed

double, treadmill speed on given training day

training.day10_treadmillincline

double, treadmill incline on given training day

training.day10_timeontreadmill

integer, time on treadmill on given training day

training.day10_posttrainlact

double, post-training lactate on given training day

training.day10_score

integer, score on given training day

training.day10_comments

character, comments on given training day

training.day11date

character, date of given training day

training.day11_days

integer, day of given training day relative to registration.d_arrive (count)

training.day11time

character, time of given training day

training.day11_treadmillspeed

double, treadmill speed on given training day

training.day11_treadmillincline

double, treadmill incline on given training day

training.day11_timeontreadmill

integer, time on treadmill on given training day

training.day11_weight

double, weight on given training day

training.day11_posttrainlact

double, post-training lactate on given training day

training.day11_score

integer, score on given training day

training.day11_comments

character, comments on given training day

training.day12date

character, date of given training day

training.day12_days

integer, day of given training day relative to registration.d_arrive (count)

training.day12time

character, time of given training day

training.day12_treadmillspeed

double, treadmill speed on given training day

training.day12_treadmillincline

double, treadmill incline on given training day

training.day12_timeontreadmill

integer, time on treadmill on given training day

training.day12_score

integer, score on given training day

training.day12_comments

character, comments on given training day

training.day13date

character, date of given training day

training.day13_days

integer, day of given training day relative to registration.d_arrive (count)

training.day13time

character, time of given training day

training.day13_treadmillspeed

double, treadmill speed on given training day

training.day13_treadmillincline

double, treadmill incline on given training day

training.day13_timeontreadmill

integer, time on treadmill on given training day

training.day13_score

integer, score on given training day

training.day13_comments

character, comments on given training day

training.day14date

character, date of given training day

training.day14_days

integer, day of given training day relative to registration.d_arrive (count)

training.day14time

character, time of given training day

training.day14_treadmillspeed

double, treadmill speed on given training day

training.day14_treadmillincline

double, treadmill incline on given training day

training.day14_timeontreadmill

integer, time on treadmill on given training day

training.day14_score

integer, score on given training day

training.day14_comments

character, comments on given training day

training.day15date

character, date of given training day

training.day15_days

integer, day of given training day relative to registration.d_arrive (count)

training.day15time

character, time of given training day

training.day15_treadmillspeed

double, treadmill speed on given training day

training.day15_treadmillincline

double, treadmill incline on given training day

training.day15_timeontreadmill

integer, time on treadmill on given training day

training.day15_posttrainlact

double, post-training lactate on given training day

training.day15_score

integer, score on given training day

training.day15_comments

character, comments on given training day

training.day16date

character, date of given training day

training.day16_days

integer, day of given training day relative to registration.d_arrive (count)

training.day16time

character, time of given training day

training.day16_treadmillspeed

double, treadmill speed on given training day

training.day16_treadmillincline

double, treadmill incline on given training day

training.day16_timeontreadmill

integer, time on treadmill on given training day

training.day16_weight

double, weight on given training day

training.day16_posttrainlact

double, post-training lactate on given training day

training.day16_score

integer, score on given training day

training.day16_comments

character, comments on given training day

training.day17date

character, date of given training day

training.day17_days

integer, day of given training day relative to registration.d_arrive (count)

training.day17time

character, time of given training day

training.day17_treadmillspeed

double, treadmill speed on given training day

training.day17_treadmillincline

double, treadmill incline on given training day

training.day17_timeontreadmill

integer, time on treadmill on given training day

training.day17_score

integer, score on given training day

training.day17_comments

character, comments on given training day

training.day18date

character, date of given training day

training.day18_days

integer, day of given training day relative to registration.d_arrive (count)

training.day18time

character, time of given training day

training.day18_treadmillspeed

double, treadmill speed on given training day

training.day18_treadmillincline

double, treadmill incline on given training day

training.day18_timeontreadmill

integer, time on treadmill on given training day

training.day18_score

integer, score on given training day

training.day18_comments

character, comments on given training day

training.day19date

character, date of given training day

training.day19_days

integer, day of given training day relative to registration.d_arrive (count)

training.day19time

character, time of given training day

training.day19_treadmillspeed

double, treadmill speed on given training day

training.day19_treadmillincline

double, treadmill incline on given training day

training.day19_timeontreadmill

integer, time on treadmill on given training day

training.day19_score

integer, score on given training day

training.day19_comments

character, comments on given training day

training.day20date

character, date of given training day

training.day20_days

integer, day of given training day relative to registration.d_arrive (count)

training.day20time

character, time of given training day

training.day20_treadmillspeed

double, treadmill speed on given training day

training.day20_treadmillincline

double, treadmill incline on given training day

training.day20_timeontreadmill

integer, time on treadmill on given training day

training.day20_posttrainlact

double, post-training lactate on given training day

training.day20_score

integer, score on given training day

training.day20_comments

character, comments on given training day

training.day21date

character, date of given training day

training.day21_days

integer, day of given training day relative to registration.d_arrive (count)

training.day21time

character, time of given training day

training.day21_treadmillspeed

double, treadmill speed on given training day

training.day21_treadmillincline

double, treadmill incline on given training day

training.day21_timeontreadmill

integer, time on treadmill on given training day

training.day21_weight

double, weight on given training day

training.day21_posttrainlact

double, post-training lactate on given training day

training.day21_score

integer, score on given training day

training.day21_comments

character, comments on given training day

training.day22date

character, date of given training day

training.day22_days

integer, day of given training day relative to registration.d_arrive (count)

training.day22time

character, time of given training day

training.day22_treadmillspeed

double, treadmill speed on given training day

training.day22_treadmillincline

double, treadmill incline on given training day

training.day22_timeontreadmill

integer, time on treadmill on given training day

training.day22_score

integer, score on given training day

training.day22_comments

character, comments on given training day

training.day23date

character, date of given training day

training.day23_days

integer, day of given training day relative to registration.d_arrive (count)

training.day23time

character, time of given training day

training.day23_treadmillspeed

double, treadmill speed on given training day

training.day23_treadmillincline

double, treadmill incline on given training day

training.day23_timeontreadmill

integer, time on treadmill on given training day

training.day23_score

integer, score on given training day

training.day23_comments

character, comments on given training day

training.day24date

character, date of given training day

training.day24_days

integer, day of given training day relative to registration.d_arrive (count)

training.day24time

character, time of given training day

training.day24_treadmillspeed

double, treadmill speed on given training day

training.day24_treadmillincline

double, treadmill incline on given training day

training.day24_timeontreadmill

integer, time on treadmill on given training day

training.day24_score

integer, score on given training day

training.day24_comments

character, comments on given training day

training.day25date

character, date of given training day

training.day25_days

integer, day of given training day relative to registration.d_arrive (count)

training.day25time

character, time of given training day

training.day25_treadmillspeed

double, treadmill speed on given training day

training.day25_treadmillincline

double, treadmill incline on given training day

training.day25_timeontreadmill

integer, time on treadmill on given training day

training.day25_posttrainlact

double, post-training lactate on given training day

training.day25_score

integer, score on given training day

training.day25_comments

character, comments on given training day

training.day26date

character, date of given training day

training.day26_days

integer, day of given training day relative to registration.d_arrive (count)

training.day26time

character, time of given training day

training.day26_treadmillspeed

double, treadmill speed on given training day

training.day26_treadmillincline

double, treadmill incline on given training day

training.day26_timeontreadmill

integer, time on treadmill on given training day

training.day26_weight

double, weight on given training day

training.day26_posttrainlact

double, post-training lactate on given training day

training.day26_score

integer, score on given training day

training.day26_comments

character, comments on given training day

training.day27date

character, date of given training day

training.day27_days

integer, day of given training day relative to registration.d_arrive (count)

training.day27time

character, time of given training day

training.day27_treadmillspeed

double, treadmill speed on given training day

training.day27_treadmillincline

double, treadmill incline on given training day

training.day27_timeontreadmill

integer, time on treadmill on given training day

training.day27_score

integer, score on given training day

training.day27_comments

character, comments on given training day

training.day28date

character, date of given training day

training.day28_days

integer, day of given training day relative to registration.d_arrive (count)

training.day28time

character, time of given training day

training.day28_treadmillspeed

double, treadmill speed on given training day

training.day28_treadmillincline

double, treadmill incline on given training day

training.day28_timeontreadmill

integer, time on treadmill on given training day

training.day28_score

integer, score on given training day

training.day28_comments

character, comments on given training day

training.day29date

character, date of given training day

training.day29_days

integer, day of given training day relative to registration.d_arrive (count)

training.day29time

character, time of given training day

training.day29_treadmillspeed

double, treadmill speed on given training day

training.day29_treadmillincline

double, treadmill incline on given training day

training.day29_timeontreadmill

integer, time on treadmill on given training day

training.day29_score

integer, score on given training day

training.day29_comments

character, comments on given training day

training.day30date

character, date of given training day

training.day30_days

integer, day of given training day relative to registration.d_arrive (count)

training.day30time

character, time of given training day

training.day30_treadmillspeed

double, treadmill speed on given training day

training.day30_treadmillincline

double, treadmill incline on given training day

training.day30_timeontreadmill

integer, time on treadmill on given training day

training.day30_posttrainlact

double, post-training lactate on given training day

training.day30_score

integer, score on given training day

training.day30_comments

character, comments on given training day

training.day31date

character, date of given training day

training.day31_days

integer, day of given training day relative to registration.d_arrive (count)

training.day31time

character, time of given training day

training.day31_treadmillspeed

double, treadmill speed on given training day

training.day31_treadmillincline

double, treadmill incline on given training day

training.day31_timeontreadmill

integer, time on treadmill on given training day

training.day31_weight

double, weight on given training day

training.day31_posttrainlact

double, post-training lactate on given training day

training.day31_score

integer, score on given training day

training.day31_comments

character, comments on given training day

training.day32date

character, date of given training day

training.day32_days

integer, day of given training day relative to registration.d_arrive (count)

training.day32time

character, time of given training day

training.day32_treadmillspeed

double, treadmill speed on given training day

training.day32_treadmillincline

double, treadmill incline on given training day

training.day32_timeontreadmill

integer, time on treadmill on given training day

training.day32_score

integer, score on given training day

training.day32_comments

character, comments on given training day

training.day33date

character, date of given training day

training.day33_days

integer, day of given training day relative to registration.d_arrive (count)

training.day33time

character, time of given training day

training.day33_treadmillspeed

double, treadmill speed on given training day

training.day33_treadmillincline

double, treadmill incline on given training day

training.day33_timeontreadmill

integer, time on treadmill on given training day

training.day33_score

integer, score on given training day

training.day33_comments

character, comments on given training day

training.day34date

character, date of given training day

training.day34_days

integer, day of given training day relative to registration.d_arrive (count)

training.day34time

character, time of given training day

training.day34_treadmillspeed

double, treadmill speed on given training day

training.day34_treadmillincline

double, treadmill incline on given training day

training.day34_timeontreadmill

integer, time on treadmill on given training day

training.day34_score

integer, score on given training day

training.day34_comments

character, comments on given training day

training.day35date

character, date of given training day

training.day35_days

integer, day of given training day relative to registration.d_arrive (count)

training.day35time

character, time of given training day

training.day35_treadmillspeed

double, treadmill speed on given training day

training.day35_treadmillincline

double, treadmill incline on given training day

training.day35_timeontreadmill

integer, time on treadmill on given training day

training.day35_posttrainlact

double, post-training lactate on given training day

training.day35_score

integer, score on given training day

training.day35_comments

character, comments on given training day

training.day36date

character, date of given training day

training.day36_days

integer, day of given training day relative to registration.d_arrive (count)

training.day36time

character, time of given training day

training.day36_weight

double, weight on given training day

training.day36_comments

character, comments on given training day

training.day37date

character, date of given training day

training.day37_days

integer, day of given training day relative to registration.d_arrive (count)

training.day37time

character, time of given training day

training.day37_treadmillspeed

double, treadmill speed on given training day

training.day37_treadmillincline

double, treadmill incline on given training day

training.day37_timeontreadmill

integer, time on treadmill on given training day

training.day37_score

integer, score on given training day

training.day37_comments

character, comments on given training day

training.day38date

character, date of given training day

training.day38_days

integer, day of given training day relative to registration.d_arrive (count)

training.day38time

character, time of given training day

training.day38_treadmillspeed

double, treadmill speed on given training day

training.day38_treadmillincline

double, treadmill incline on given training day

training.day38_timeontreadmill

integer, time on treadmill on given training day

training.day38_score

integer, score on given training day

training.day38_comments

character, comments on given training day

training.day39date

character, date of given training day

training.day39_days

integer, day of given training day relative to registration.d_arrive (count)

training.day39time

character, time of given training day

training.day39_treadmillspeed

double, treadmill speed on given training day

training.day39_treadmillincline

double, treadmill incline on given training day

training.day39_timeontreadmill

integer, time on treadmill on given training day

training.day39_posttrainlact

double, post-training lactate on given training day

training.day39_score

integer, score on given training day

training.day39_comments

character, comments on given training day

training.day40date

character, date of given training day

training.day40_days

integer, day of given training day relative to registration.d_arrive (count)

training.day40time

character, time of given training day

training.day40_treadmillspeed

double, treadmill speed on given training day

training.day40_treadmillincline

double, treadmill incline on given training day

training.day40_timeontreadmill

integer, time on treadmill on given training day

training.day40_posttrainlact

double, post-training lactate on given training day

training.day40_score

integer, score on given training day

training.day40_comments

character, comments on given training day

vo2.max.test.d_visit_1

character, date first VO2 max test was completed

vo2.max.test.d_visit_2

character, date second VO2 max test was completed

vo2.max.test.days_visit_1

integer, day first VO2 max test was completed relative to registration.d_arrive (count)

vo2.max.test.days_visit_2

integer, day second VO2 max test was completed relative to registration.d_arrive (count)

vo2.max.test.staffid_1

integer, first VO2 max test staff ID

vo2.max.test.staffid_2

integer, second VO2 max test staff ID

vo2.max.test.siteid_1

character, first VO2 max test site

vo2.max.test.siteid_2

character, second VO2 max test site

vo2.max.test.visitcode_1

character, visit code for first VO2 max test

vo2.max.test.visitcode_2

character, visit code for second VO2 max test

vo2.max.test.formname_1

character, form name for first VO2 max test

vo2.max.test.formname_2

character, form name for second VO2 max test

vo2.max.test.versionnbr_1

integer, form version number for first VO2 max test

vo2.max.test.versionnbr_2

integer, form version number for second VO2 max test

vo2.max.test.d_vo2_1

character, date of first VO2 max test

vo2.max.test.d_vo2_2

character, date of second VO2 max test

vo2.max.test.days_vo2_1

integer, date of first VO2 max test relative to registration.d_arrive (count)

vo2.max.test.days_vo2_2

integer, date of second VO2 max test relative to registration.d_arrive (count)

vo2.max.test.t_vo2_1

character, start time of first VO2 max test?

vo2.max.test.t_vo2_2

character, start time of second VO2 max test?

vo2.max.test.blactate_begin_1

double, blood lactate at the beginning of exercise for first VO2 max test

vo2.max.test.blactate_begin_2

double, blood lactate at the beginning of exercise for second VO2 max test

vo2.max.test.vo2_max_1

double, VO2 at max for first VO2 max test

vo2.max.test.vo2_max_2

double, VO2 at max for second VO2 max test

vo2.max.test.vco2_max_1

double, VCO2 at max for first VO2 max test

vo2.max.test.vco2_max_2

double, VCO2 at max for second VO2 max test

vo2.max.test.rer_max_1

double, RER at max for first VO2 max test

vo2.max.test.rer_max_2

double, RER at max for second VO2 max test

vo2.max.test.speed_max_1

double, max speed of running for first VO2 max test

vo2.max.test.speed_max_2

double, max speed of running for second VO2 max test

vo2.max.test.t_complete_1

character, time of completion for first VO2 max test

vo2.max.test.t_complete_2

character, time of completion for second VO2 max test

vo2.max.test.blactate_end_1

double, blood lactate at end of exercise for first VO2 max test

vo2.max.test.blactate_end_2

double, blood lactate at end of exercise for second VO2 max test

vo2.max.test.vo2_comments_1

character, comments about first VO2 max test

vo2.max.test.vo2_comments_2

character, comments about second VO2 max test

nmr.testing.d_visit_1

character, date first NMR test was completed

nmr.testing.d_visit_2

character, date second NMR test was completed

nmr.testing.days_visit_1

integer, day first NMR test was completed relative to registration.d_arrive (count)

nmr.testing.days_visit_2

integer, day second NMR test was completed relative to registration.d_arrive (count)

nmr.testing.staffid_1

integer, first NMR test staff ID

nmr.testing.staffid_2

integer, second NMR test staff ID

nmr.testing.siteid_1

character, first NMR test site

nmr.testing.siteid_2

character, second NMR test site

nmr.testing.visitcode_1

character, visit code for first NMR test

nmr.testing.visitcode_2

character, visit code for second NMR test

nmr.testing.formname_1

character, form name for first NMR test

nmr.testing.formname_2

character, form name for second NMR test

nmr.testing.versionnbr_1

integer, form version number for first NMR test

nmr.testing.versionnbr_2

integer, form version number for second NMR test

nmr.testing.d_nmr_1

character, date data was collected for the first NMR test

nmr.testing.d_nmr_2

character, date data was collected for the second NMR test

nmr.testing.days_nmr_1

integer, day data was collected for the first NMR test relative to registration.d_arrive (count)

nmr.testing.days_nmr_2

integer, day data was collected for the second NMR test relative to registration.d_arrive (count)

nmr.testing.nmr_weight_1

double, rat's body weight for first NMR test

nmr.testing.nmr_weight_2

double, rat's body weight for second NMR test

nmr.testing.nmr_fat_1

double, rat's percent body fat for first NMR test

nmr.testing.nmr_fat_2

double, rat's percent body fat for second NMR test

nmr.testing.nmr_lean_1

double, rat's percent lean for first NMR test

nmr.testing.nmr_lean_2

double, rat's percent lean for second NMR test

nmr.testing.nmr_fluid_1

double, rat's percent fluid for first NMR test

nmr.testing.nmr_fluid_2

double, rat's percent fluid for second NMR test

nmr.testing.nmr_comments_2

character, comments about second NMR test

terminal.weight.bw

double, terminal body weight (g)

terminal.weight.lg

integer, terminal lateral gastrocnemius weight (mg)

terminal.weight.mg

integer, terminal medial gastrocnemius weight (mg)

terminal.weight.pl

integer, terminal plantaris weight (mg)

terminal.weight.sol

integer, terminal soleus weight (mg)

specimen.collection.d_visit

character, date of specimen collection

specimen.collection.days_visit

integer, day of specimen collection relative to registration.d_arrive (count)

specimen.collection.staffid

integer, specimen collection staff ID

specimen.collection.siteid

character, specimen collection site

specimen.collection.visitcode

character, specimen collection site visit code

specimen.collection.formname

character, specimen collection form name

specimen.collection.versionnbr

integer, specimen collection form version number

specimen.collection.anesthesiaid

integer, anesthesia administrator ID

specimen.collection.t_anesthesia

character, anesthesia administration time

specimen.collection.bloodtype

character, blood type

specimen.collection.bloodtube

character, blood tube

specimen.collection.bloodcomplete

character, blood collection complete

specimen.collection.bloodtechid

character, blood collection technician ID

specimen.collection.t_bloodstart

character, blood collection start time

specimen.collection.t_bloodstop

character, blood collection stop time

specimen.collection.t_edtafill

character, EDTA fill time

specimen.collection.uterustype

character, uterus type

specimen.collection.uterustechid

character, uterus technician ID

specimen.collection.uteruscomplete

character, uterus complete

specimen.collection.t_death

character, time of death

specimen.collection.deathtype

character, type of death

specimen.processing.versionnbr

integer, specimen processing form version number

specimen.processing.formname

character, specimen processing form name

specimen.processing.visitcode

character, specimen processing visit code

specimen.processing.siteid

character, specimen processing site

specimen.processing.sampletypedescription

character, specimen tissue type

specimen.processing.samplenumber

integer, specimen sample number

specimen.processing.timepoint

integer, specimen time point

specimen.processing.aliquotdescription

character, specimen aliquot tube

specimen.processing.volume

character, specimen aliquot volume

specimen.processing.hemolyzed

character, specimen hemolyzed?

specimen.processing.t_collection

character, specimen collection time

specimen.processing.t_edtaspin

character, EDTA spin time

specimen.processing.t_freeze

character, specimen freeze time

specimen.processing.techid

character, specimen processing tech ID

calculated.variables.pct_body_fat_change

double, percent body fat change after training: nmr.testing.nmr_fat_2 - nmr.testing.nmr_fat_1

calculated.variables.pct_body_lean_change

double, percent body lean change after training: nmr.testing.nmr_lean_2 - nmr.testing.nmr_lean_1

calculated.variables.pct_body_fluid_change

double, percent body fluid change after training: nmr.testing.nmr_fluid_2 - nmr.testing.nmr_fluid_1

calculated.variables.lactate_change_dueto_train

double, lactate change due to training: vo2.max.test.blactate_end_2 - vo2.max.test.blactate_begin_1

calculated.variables.vo2_max_change

double, change in VO2 max after training: vo2.max.test.vo2_max_2 - vo2.max.test.vo2_max_1

calculated.variables.coll_time_train

integer, sample collection time after training (seconds): specimen collection date and time - last time trained

calculated.variables.deathtime_after_train

integer, time of death after training (seconds): death date and time - last time trained

calculated.variables.frozetime_after_train

integer, time of sample freezing after training (seconds): specimen collection date and freeze time - last time trained

biclabeldata.shiptositeid

character, Chemical Analysis Site (CAS), i.e. the omics site)

sacrificetime

character, sacrifice time point in weeks, one of "1w", "2w", "4w", "8w"

intervention

character, intervention, one of "training", "control"

group

character, intervention group, one of "1w", "2w", "4w", "8w", "control"

sex

character, one of 'male' or 'female'

time_to_freeze

integer, sample time to freeze, calculated.variables.frozetime_after_train - calculated.variables.deathtime_after_train

tissue_code_no

character, BIC tissue code. See MotrpacBicQC::bic_animal_tissue_code.

tissue

character, tissue abbreviation, one of TISSUE_ABBREV