Repeated feature info

Documentation of 91 IMMUNO and METAB features that were measured on multiple platforms and have multiple differential analysis results. New unique identifiers are defined. These new unique identifiers were used in the graphical analysis.

Usage

REPEATED_FEATURES

Format

A data frame with 825 rows and 10 variables:

feature_ID

character, MoTrPAC feature identifier

assay

character, assay abbreviation, one of ASSAY_ABBREV

tissue

character, tissue abbreviation, one of TISSUE_ABBREV

dataset

character, for immunoassay and metabolomics features, this variable specifies the immune panel (rat-myokine, rat-pituitary, rat-mag27plex, rat-metabolic, ADIPONECTIN, SERPIN-E) or metabolomics platform (metab-u-ionpneg, metab-u-lrpneg, metab-u-lrppos, metab-u-hilicpos, metab-u-rpneg, metab-u-rppos, metab-t-amines, metab-t-oxylipneg, metab-t-tca, metab-t-nuc, metab-t-acoa, metab-t-ka) the feature was measured in. 'meta-reg' specifies results from the metabolomics meta-regression for repeated features

selection_fdr

double, adjusted training p-value used to select training-regulated analytes. P-values are IHW-adjusted across all datasets within a given assay with tissue as a covariate.

metabolite_refmet

character, RefMet name of metabolite

feature

character, duplicated feature in the format ASSAY_ABBREV;TISSUE_ABBREV;[feature_ID]

new_feature_ID

character, new unique feature_ID with dataset prepended

new_feature

character, new unique feature in the format ASSAY_ABBREV;TISSUE_ABBREV;[new_feature_ID]

Details

91 IMMUNO and METAB features were measured on multiple platforms and have multiple differential analysis results. In order to distinguish between the different sets of results, these feature_IDs were prepended with dataset. For example, "BDNF" is separated into "rat-myokine:BDNF" and "rat-pituitary:BDNF". For completeness, both the modified and unmodified feature_IDs are included in both the feature-to-gene map and universe lists. Note that the graphical analysis uses the modified feature_IDs; differential analysis results use the unmodified feature_IDs.