repfdr results

Raw repfdr::repfdr() results from which the graphical state assignments were determined

Usage

REPFDR_RES

Format

List:

repfdr_em_res

a list with repfdr's EM results

repfdr_clusters

repfdr's configurations

repfdr_clusters_str

repfdr's configurations, string representation

repfdr_clusters_pi

configuration's inferred priors

Details

repfdr::repfdr() is an algorithm suggested by Yekutieli and Heller in 2014 (Bioinformatics) for analysis of p-values or z-scores from different resources. It is based on the assumption that z-scores from each resource (a time point from a specific sex in our case) is either positive, null, or negative. Then the algorithm internally learns the mixture distribution in each resource and the dependencies among them. In the process, a simplifying assumption is made about conditional independence between the z-scores given their state (so for example, two high scores from weeks 4 and 8 in males are independent given that we know that they are positive, non-null cases).

Our analysis pipeline is as follows: we run repfdr::repfdr() on the z-score data matrix of all training-regulated features at 5% IWH FDR (see REPFDR_INPUTS). Then we use the output to extract the analytes of each state in each time point. Here a state means one of (male up, male null, male down) x (female up, female null, female down) for each time point. Once these are selected we call them the node_sets. Then, for each pair of node (x,y) such that y is from a time point that is adjacent and after x (e.g., x is a node from week 4 and y is a node from week 8), we define their edge set as the intersection of their analytes.

Reproduce our analysis with MotrpacRatTraining6mo::bayesian_graphical_clustering() and MotrpacRatTraining6mo::repfdr_wrapper().