Differential analysis of RNA-seq datasets

Timewise summary statistics and training FDR from differential analysis (DA) that tests the effect of training on the expression of each gene within each sex. One data frame per tissue.

Usage

TRNSCRPT_BLOOD_DA

TRNSCRPT_HIPPOC_DA

TRNSCRPT_CORTEX_DA

TRNSCRPT_HYPOTH_DA

TRNSCRPT_SKMGN_DA

TRNSCRPT_SKMVL_DA

TRNSCRPT_HEART_DA

TRNSCRPT_KIDNEY_DA

TRNSCRPT_ADRNL_DA

TRNSCRPT_COLON_DA

TRNSCRPT_SPLEEN_DA

TRNSCRPT_TESTES_DA

TRNSCRPT_OVARY_DA

TRNSCRPT_VENACV_DA

TRNSCRPT_LUNG_DA

TRNSCRPT_SMLINT_DA

TRNSCRPT_LIVER_DA

TRNSCRPT_BAT_DA

TRNSCRPT_WATSC_DA

Format

A data frame with 22 variables:

feature

character, unique feature identifier in the format 'ASSAY_ABBREV;TISSUE_ABBREV;feature_ID' only for training-regulated features at 5% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See REPEATED_FEATURES for details.

assay

character, assay abbreviation, one of ASSAY_ABBREV

assay_code

character, assay code used in data release. See MotrpacBicQC::assay_codes.

tissue

character, tissue abbreviation, one of TISSUE_ABBREV

tissue_code

character, tissue code used in data release. See MotrpacBicQC::bic_animal_tissue_code.

feature_ID

character, MoTrPAC feature identifier

sex

character, one of 'male' or 'female'

comparison_group

character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'

p_value

double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals

adj_p_value

double, adjusted p-value from 'p_value' column. P-values are BY-adjusted across all datasets within a given assay/ome.

logFC

double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals

logFC_se

double, standard error of the log fold-change

shrunk_logFC

double, log fold-change with shrinkage applied

shrunk_logFC_se

double, standard error of the shrunken log fold-change

zscore

double, z statistic

covariates

character, comma-separated list of adjustment variables or NA

removed_samples

character, comma-separated list of vial labels excluded from differential analysis or NA

comparison_average_intensity

double, average intensity among the replicates in 'comparison_group'

comparison_average_intensity_se

double, standard error of 'comparison_average_intensity'

reference_average_intensity

double, average intensity among the replicates in the group of sex-matched sedentary control animals

reference_average_intensity_se

double, standard error of 'reference_average_intensity'

selection_fdr

double, adjusted training p-value used to select training-regulated analytes. P-values are IHW-adjusted across all datasets within a given assay with tissue as a covariate.

An object of class data.frame with 132864 rows and 22 columns.

An object of class data.frame with 131544 rows and 22 columns.

An object of class data.frame with 135640 rows and 22 columns.

An object of class data.frame with 110192 rows and 22 columns.

An object of class data.frame with 111784 rows and 22 columns.

An object of class data.frame with 115560 rows and 22 columns.

An object of class data.frame with 127848 rows and 22 columns.

An object of class data.frame with 126472 rows and 22 columns.

An object of class data.frame with 132448 rows and 22 columns.

An object of class data.frame with 130552 rows and 22 columns.

An object of class data.frame with 69436 rows and 22 columns.

An object of class data.frame with 68140 rows and 22 columns.

An object of class data.frame with 98028 rows and 22 columns.

An object of class data.frame with 132040 rows and 22 columns.

An object of class data.frame with 131288 rows and 22 columns.

An object of class data.frame with 115496 rows and 22 columns.

An object of class data.frame with 129232 rows and 22 columns.

An object of class data.frame with 134112 rows and 22 columns.

Details

Reproduce our analysis with MotrpacRatTraining6mo::transcript_timewise_da() and MotrpacRatTraining6mo::transcript_training_da().