Metabolite feature IDs and metadata

Mapping between various metabolite feature identifiers used at different stages of data processing. Also includes some feature metadata

Usage

METAB_FEATURE_ID_MAP

Format

A data frame with 14420 rows and 13 variables:

tissue

character, tissue abbreviation, one of TISSUE_ABBREV

dataset

character, metabolomics platform (metab-u-ionpneg, metab-u-lrpneg, metab-u-lrppos, metab-u-hilicpos, metab-u-rpneg, metab-u-rppos, metab-t-amines, metab-t-oxylipneg, metab-t-tca, metab-t-nuc, metab-t-acoa, metab-t-ka) the feature was measured in. 'meta-reg' specifies results from the metabolomics meta-regression for repeated features.

metabolite_name

character, name of metabolite as appears in the Chemical Analysis Site's data

metabolite_refmet

character, RefMet name of metabolite

feature_ID_sample_data

character, feature_ID used in the sample-level data, e.g. METAB_NORM_DATA_FLAT

feature_ID_da

character, feature_ID used in the standard differential analysis results tables, e.g. METAB_BAT_DA

feature_ID_metareg

character, feature_ID used in the meta-regression results tables, e.g. METAB_BAT_DA_METAREG

dataset_metareg

character, version of the differential analysis results retained after meta-regression, either dataset or 'meta-reg'

feature

character, unique feature identifier in the format 'ASSAY_ABBREV;TISSUE_ABBREV;feature_ID' only for training-regulated features at 5% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See REPEATED_FEATURES for details.

rt

double, retention time

mz

double, mass over charge

neutral_mass

double, neutral mass

formula

character, chemical formula

Details

metabolite_name is always equal to feature_ID_sample_data and feature_ID_da.

feature_ID_metareg is different than metabolite_name when the combined results from multiple measurements of the same metabolite are reported in the meta-regression differential analysis results, in which case feature_ID_metareg is metabolite_refmet. 112 rows (57 unique metabolites in 9 tissues) do not have corresponding meta-regression results, in which case feature_ID_metareg is NA.

feature is only non-NA for training-regulated features at 5% IHW FDR (see TRAINING_REGULATED_FEATURES). Each non-NA value of feature is unique. In the case of repeated measurements, this means dataset was added to metabolite_name. See REPEATED_FEATURES for more details.

Note that RefMet IDs are frequently updated. This table provides the version of the RefMet IDs at the time that these data were generated.