Differential analysis of proteomics datasets

Timewise summary statistics and training FDR from differential analysis (DA) that tests the effect of training on each proteomics feature (protein, phosphosite, acetylsite, ubiquitlsite) within each sex. One data frame per data type per tissue.

Usage

PROT_CORTEX_DA

PROT_SKMGN_DA

PROT_HEART_DA

PROT_KIDNEY_DA

PROT_LUNG_DA

PROT_LIVER_DA

PROT_WATSC_DA

PHOSPHO_CORTEX_DA

PHOSPHO_SKMGN_DA

PHOSPHO_HEART_DA

PHOSPHO_KIDNEY_DA

PHOSPHO_LUNG_DA

PHOSPHO_LIVER_DA

PHOSPHO_WATSC_DA

ACETYL_HEART_DA

ACETYL_LIVER_DA

UBIQ_HEART_DA

UBIQ_LIVER_DA

Format

A data frame 18 variables:

feature

character, unique feature identifier in the format 'ASSAY_ABBREV;TISSUE_ABBREV;feature_ID' only for training-regulated features at 5% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See REPEATED_FEATURES for details.

assay

character, assay abbreviation, one of ASSAY_ABBREV

assay_code

character, assay code used in data release. See MotrpacBicQC::assay_codes.

tissue

character, tissue abbreviation, one of TISSUE_ABBREV

tissue_code

character, tissue code used in data release. See MotrpacBicQC::bic_animal_tissue_code.

feature_ID

character, MoTrPAC feature identifier

sex

character, one of 'male' or 'female'

comparison_group

character, time point of trained animals compared to the sex-matched sedentary control animals, one of '1w', '2w', '4w', '8w'

p_value

double, unadjusted p-value for the difference between 'comparison_group' and the group of sex-matched sedentary control animals

adj_p_value

double, adjusted p-value from 'p_value' column. P-values are BY-adjusted across all datasets within a given assay/ome.

logFC

double, log fold-change where the numerator is 'comparison_group', e.g., '1w', and the denominator is the group of sex-matched sedentary control animals

logFC_se

double, standard error of the log fold-change

tscore

double, t statistic

covariates

character, comma-separated list of adjustment variables or NA

numNAs

integer, number of missing values in the current 'comparison_group' and sex-matched sedentary control animals

comparison_average_intensity

double, average intensity among the replicates in 'comparison_group'

reference_average_intensity

double, average intensity among the replicates in the group of sex-matched sedentary control animals

selection_fdr

double, adjusted training p-value used to select training-regulated analytes. P-values are IHW-adjusted across all datasets within a given assay with tissue as a covariate.

An object of class data.frame with 47992 rows and 18 columns.

An object of class data.frame with 73472 rows and 18 columns.

An object of class data.frame with 78816 rows and 18 columns.

An object of class data.frame with 87016 rows and 18 columns.

An object of class data.frame with 78328 rows and 18 columns.

An object of class data.frame with 79712 rows and 18 columns.

An object of class data.frame with 372016 rows and 18 columns.

An object of class data.frame with 237208 rows and 18 columns.

An object of class data.frame with 321664 rows and 18 columns.

An object of class data.frame with 241152 rows and 18 columns.

An object of class data.frame with 381288 rows and 18 columns.

An object of class data.frame with 348536 rows and 18 columns.

An object of class data.frame with 242432 rows and 18 columns.

An object of class data.frame with 41704 rows and 18 columns.

An object of class data.frame with 78000 rows and 18 columns.

An object of class data.frame with 56624 rows and 18 columns.

An object of class data.frame with 74752 rows and 18 columns.

Details

Reproduce our analysis with MotrpacRatTraining6mo::proteomics_timewise_da() and MotrpacRatTraining6mo::proteomics_training_da().