Training differential analysis results

Training summary statistics from differential analysis (DA) that tests the effect of training on each feature within each sex. One data frame per data type and tissue.

Format

A data frame with up to 47 variables:

assay

character, assay code used in data release. See MotrpacBicQC::assay_codes.

tissue

character, tissue code used in data release. See MotrpacBicQC::bic_animal_tissue_code.

feature_ID

character, MoTrPAC feature identifier

removed_samples

character, comma-separated list of vial labels excluded from differential analysis or NA

removed_samples_male

character, comma-separated list of male outliers (vial labels) removed from differential analysis

removed_samples_female

character, comma-separated list of female outliers (vial labels) removed from differential analysis

dataset

character, specific dataset

dataset_male

character, specific dataset

dataset_female

character, specific dataset

panel

character, LUMINEX panel for immunoassay analytes

full_model

character, full model used for differential analysis

full_model_male

character, full model used for differential analysis in males

full_model_female

character, full model used for differential analysis in females

reduced_model

character, effective reduced model used for differential analysis

reduced_model_male

character, effective reduced model used for differential analysis in males

reduced_model_female

character, effective reduced model used for differential analysis in females

covariates

character, comma-separated list of adjustment variables

covariates_male

character, comma-separated list of adjustment variables used for males

covariates_female

character, comma-separated list of adjustment variables used for females

fscore_male

double, F-statistic for males

fscore_female

double, F-statistic for females

lrt

double, likelihood ratio test statistic when only one sex exists

lrt_male

double, likelihood ratio test statistic for males

lrt_female

double, likelihood ratio test statistic for females

p_value_male

double, nominal p-value for males

p_value_female

character, nominal p-value for females

adj_p_value_male

double, IHW-adjusted p-value for males

adj_p_value_female

double, IHW-adjusted p-value for females

p_value

double, combined male and female nominal p-value using the sum of logs

adj_p_value

double, IHW-adjusted combined p-value; value provided by selection_fdr in the timewise summary statistics

tissue_abbreviation

character, tissue abbreviation, one of TISSUE_ABBREV

Chr

character, chromosome where feature is located. METHYL only.

Locus

character, name of feature. METHYL only.

EntrezID

integer, Entrez ID of corresponding gene. METHYL only.

Symbol

character, gene symbol of corresponding gene. METHYL only.

is_targeted

logical, whether metabolite is targeted. Metabolomics only.

site

character, Chemical Analysis Site (CAS) name. Metabolomics only.

metabolite

character, name of metabolite as appears in the CAS's data. Metabolomics only.

metabolite_refmet

character, RefMet name of metabolite. Metabolomics only.

groups_tested_male

character, timepoints used to perform the F-test in males. Some tissues or assays are missing timepoints. Metabolomics only.

groups_tested_female

character, timepoints used to perform the F-test in females. Some tissues or assays are missing timepoints. Metabolomics only.

mz

double, mass over charge. Metabolomics only.

rt

double, retention time. Metabolomics only.

neutral_mass

numeric, neutral mass. Metabolomics only.

cv

double, feature coffeicient of variation in the dataset. Metabolomics only.

meta_reg_het_p

double, heterogeneity p-value from meta-regression. A smaller p-value indicates more disagreement between platforms. Metabolomics meta-regression only.

original_ftest_ps

character, comma-separated list of original F-test p-values. Metabolomics meta-regression only.

Details

While the more commonly-used timewise summary statistics are available directly in this package (e.g., TRNSCRPT_DA, PROTEOME_DA, IMMUNO_DA, METAB_DA, METAB_DA_METAREG, TRAINING_REGULATED_FEATURES, REPFDR_INPUTS), training summary statistics are available for download from Google Cloud Storage. Use MotrpacRatTraining6mo::load_training_da() to download results for a given tissue and ome. Note that the adjusted p-value from the training differential analysis is already included in the timewise summary statistics as selection_fdr.

Training differential analysis was performed using either F-tests or likelihood ratio tests depending on the ome. Find the functions used to generate these results listed here. See details in the supplementary methods of the manuscript.