Training-regulated features

Differential analysis results for training-regulated features at 5% FDR

Usage

TRAINING_REGULATED_FEATURES

Format

A data frame with 279002 rows and 18 variables:

feature

character, unique feature identifier in the format 'ASSAY_ABBREV;TISSUE_ABBREV;feature_ID' only for training-regulated features at 5% IHW FDR. For redundant differential features, 'feature_ID' is prepended with the specific platform to make unique identifiers. See REPEATED_FEATURES for details.

assay

character, assay abbreviation, one of ASSAY_ABBREV

assay_code

character, assay code used in data release. See MotrpacBicQC::assay_codes.

tissue

character, tissue abbreviation, one of TISSUE_ABBREV

tissue_code

character, tissue code used in data release. See MotrpacBicQC::bic_animal_tissue_code.

feature_ID

character, MoTrPAC feature identifier

non_redundant_feature_ID

character, non-redundant feature identifier defined for feature_IDs with measurements from multiple platforms. feature uses non_redundant_feature_ID when available and feature_ID otherwise. See REPEATED_FEATURES for more details.

platform

character, for immunoassay and metabolomics features, this variable specifies the immune panel (rat-myokine, rat-pituitary, rat-mag27plex, rat-metabolic, ADIPONECTIN, SERPIN-E) or metabolomics platform (metab-u-ionpneg, metab-u-lrpneg, metab-u-lrppos, metab-u-hilicpos, metab-u-rpneg, metab-u-rppos, metab-t-amines, metab-t-oxylipneg, metab-t-tca, metab-t-nuc, metab-t-acoa, metab-t-ka) the feature was measured in. 'meta-reg' specifies results from the metabolomics meta-regression for repeated features

sex

character, one of 'male' or 'female'

training_group

character, training time point in weeks, one of "1w", "2w", "4w", "8w". Corresponds to comparison_group in the *_DA tables.

timewise_logFC

double, log fold-change of the training group specified by sex and training_group (e.g., 1-week females) relative to the sex-matched sedentary controls. Corresponds to logFC in the *_DA tables.

timewise_logFC_se

double, standard error of timewise_logFC. Corresponds to logFC_se in the *_DA tables.

timewise_p_value

double, timewise p-value corresponding to the contrast between the training group (e.g., 1-week females) and the sex-matched sedentary controls. Corresponds to p_value in the *_DA tables.

timewise_zscore

double, smoothed z-score corresponding to timewise_p_value used for graphical clustering. Corresponds to zscore or tscore in the *_DA tables.

meta_reg_het_p

double, for metabolites with multiple measurements, the meta-regression heterogeneity p-value, where a smaller p-value indicates more disagreement between platforms. One value per feature (not per training group).

meta_reg_pvalue

double, for metabolites with multiple measurements, the meta-regression p-value. One value per feature (not per training group).

training_p_value

double, combined training p-value determined from F-tests or LRTs comparing a full model with ome-specific technical covariates and training group as a factor variable against a reduced model with only technical covariates. Sex-specific training p-values were merged using the sum of logs. One value per features (not per training group)

training_q

double, IHW FDR-adjusted training_p_value; training-regulated features were defined by training_q < 0.05. Corresponds to selection_fdr in the *_DA tables.